wip: js cleanup
