+++ elixir/web-ng/assets/js/lib/god_view/layout_methods.js
@@ -1,0 +1,10 @@
+import {godViewLayoutAnimationMethods} from "./layout_animation_methods"
+import {godViewLayoutClusterMethods} from "./layout_cluster_methods"
+import {godViewLayoutTopologyMethods} from "./layout_topology_methods"
+
+export const godViewLayoutMethods = Object.assign(
+  {},
+  godViewLayoutClusterMethods,
+  godViewLayoutAnimationMethods,
+  godViewLayoutTopologyMethods,
+)
