+++ elixir/web-ng/assets/js/lib/god_view/rendering_style_methods.js
@@ -1,0 +1,10 @@
+import {godViewRenderingStyleEdgeMethods} from "./rendering_style_edge_methods"
+import {godViewRenderingStyleNodeMethods} from "./rendering_style_node_methods"
+import {godViewRenderingStylePipelineMethods} from "./rendering_style_pipeline_methods"
+
+export const godViewRenderingStyleMethods = Object.assign(
+  {},
+  godViewRenderingStyleNodeMethods,
+  godViewRenderingStyleEdgeMethods,
+  godViewRenderingStylePipelineMethods,
+)
